requirements.txt
flask==2.3.3
google-generativeai==0.3.2
python-dotenv==1.0.0
Pillow==10.0.0
python-multipart==0.0.6
